Updates the instruction for oh-my-zsh
Readme mention there's no dependencies, but the theme requires that some submodules are fetched. This PR intends to make it explicit, as this is useful for oh-my-zsh users. Also the instructions in the oh-my-zsh wiki are not clear enough when the theme is a bit more complex than a single file. The git checkout instructions and theme set-up allow to clone several themes in their own directory.